src/auth/requirements.txt
aniso8601==8.0.0
certifi==2019.11.28
cryptography==2.8
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
redis==3.3.11
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.12
urllib3==1.25.7
Werkzeug==0.16.0
